Struct gazetta::markdown::Markdown

pub struct Markdown<'a> {
    // some fields omitted
}

Markdown renderer

Methods

impl<'a> Markdown<'a>

fn new(data: &'a str, base: &'a str) -> Markdown<'a>

Create a new markdown renderer.

data should contain the markdown to be rendered and base should specify a relative url prefix (for relative links and images).

Note: base will only affect markdown links and images, not inline html ones.