Struct gazetta::Entry

pub struct Entry<EntryMeta> where EntryMeta: Meta {
    pub title: String,
    pub date: Option<Date>,
    pub index: Option<Index>,
    pub cc: Vec<String>,
    pub meta: EntryMeta,
    pub name: String,
    pub content: ContentSource,
}

An entry in the website.

Note: Pages do not correspond one-to-one to Entries (due to pagination).

Fields

title

The entry's title.

This is separate from the general metadata for sorting and linking. All entries should have titles.

date

The entry's date.

This is separate from the general metadata for sorting. Many entries will have dates.

index

The entries index options (if specified).

cc

Indices into which this entry should be linked.

meta

Extra metadata.

name

The entry name.

content

The content

Methods

impl<EntryMeta> Entry<EntryMeta> where EntryMeta: Meta

fn from_file(full_path: PathBuf, name: String) -> Result<Self, SourceError>