Struct gazetta::view::Site

pub struct Site<'a, G> where G: Gazetta + 'a, G::SiteMeta: 'a {
    pub title: &'a str,
    pub stylesheets: Option<&'a str>,
    pub javascript: Option<&'a str>,
    pub icon: Option<&'a str>,
    pub meta: &'a G::SiteMeta,
}

A "website"

You should include this view in your websites "head". It renders into script, stylesheet, icon tags metadata tags.

html! {
    html {
        head {
            // ...
            : site;
        }
    }
}

Fields

title

The website's title

stylesheets

The concatinated stylesheets.

javascript

The concatinated javascript.

icon

The icon.

meta

Extra metadata specified in the Source.